Read data hunks in chunks
Some makeself archives are very big (e.g., game installers can be 4 GB or more).
It's very taxing on the machine to read the whole data hunk into memory before
writing it to a separate file.

Therefore change logic to read and write each data hunk 1 MB at a time. This
makes the script a lot lighter in terms of memory usage, and extracting a 4 GB
makeself archive becomes feasible.